@@ -1,6 +1,6 @@
 {
   "name": "ngrx-store-sync",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "State and storage syncing for @ngrx/store",
   "repository": {
     "type": "git",
@@ -29,41 +29,58 @@
     "lint": "ng lint",
     "e2e": "ng e2e",
     "build_lib": "ng build ngrx-store-sync",
+    "build_dist": "npm run clean && ng build ngrx-store-sync --configuration production",
     "npm_pack": "cd dist/ngrx-store-sync && npm pack",
-    "package": "npm run build_lib && npm run npm_pack",
+    "package": "npm run build_dist && npm run npm_pack",
     "test:ci": "ng test --watch=false --browsers=ChromeHeadless",
-    "build:ci": "npm run clean && npm run build"
+    "build:ci": "npm run clean && npm run build",
+    "prepare": "husky install"
+  },
+  "engines": {
+    "node": ">= 12.18.3",
+    "yarn": ">= 1.22.0"
+  },
+  "dependencies": {
+    "@angular/compiler": "~12.1.1",
+    "@angular/core": "~12.1.1",
+    "tslib": "^2.1.0"
   },
   "devDependencies": {
-    "@angular/core": "~8.2.4",
-    "@angular-devkit/build-angular": "~0.803.2",
-    "@angular-devkit/build-ng-packagr": "~0.803.2",
-    "@angular/cli": "~8.3.2",
-    "@angular/compiler": "~8.2.4",
-    "@angular/compiler-cli": "~8.2.4",
-    "@angular/language-service": "~8.2.4",
-    "@ngrx/effects": "^8.3.0",
-    "@ngrx/store": "^8.3.0",
-    "@types/jasmine": "~3.3.8",
+    "@angular-devkit/build-angular": "~12.1.1",
+    "@angular-eslint/builder": "12.2.0",
+    "@angular-eslint/eslint-plugin": "12.2.0",
+    "@angular-eslint/eslint-plugin-template": "12.2.0",
+    "@angular-eslint/schematics": "12.2.0",
+    "@angular-eslint/template-parser": "12.2.0",
+    "@angular/cli": "~12.1.1",
+    "@angular/compiler-cli": "~12.1.1",
+    "@angular/language-service": "~12.1.1",
+    "@ngrx/effects": "^12.2.0",
+    "@ngrx/store": "^12.2.0",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.0.0",
-    "deepmerge": "^3.2.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "ng-packagr": "^5.4.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tsickle": "^0.37.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.5.3",
-    "rxjs": "~6.4.0",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "@types/node": "^12.11.1",
+    "@typescript-eslint/eslint-plugin": "^4.23.0",
+    "@typescript-eslint/parser": "^4.23.0",
+    "deepmerge": "^4.2.2",
+    "eslint": "^7.26.0",
+    "husky": "^7.0.1",
+    "jasmine-core": "~3.7.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.3.4",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "lint-staged": "^11.0.0",
+    "ng-packagr": "^12.1.0",
+    "protractor": "~7.0.0",
+    "rxjs": "~6.6.7",
+    "ts-node": "^9.1.1",
+    "typescript": "~4.3.5",
+    "zone.js": "~0.11.4"
+  },
+  "lint-staged": {
+    "*.ts": "eslint --cache --fix --max-warnings=0"
   }
 }
